主module大改，基本完成，其他后台大调整，暂角色授权待完成

+++ src/main/java/com/ssm/admin/view/RecursionMenuVo.java
@@ -14,17 +14,17 @@ public class RecursionMenuVo {
 
     private String icon;
 
-    private short seq;
+    private Integer seq;
 
     private boolean status;
 
     private Date createTime;
 
-    public short getSeq() {
+    public Integer getSeq() {
         return seq;
     }
 
-    public void setSeq(short seq) {
+    public void setSeq(Integer seq) {
         this.seq = seq;
     }
